progress in building component BottomInfoPanel7-9-23-9h15
