+++ frontend/src/components/BottomInfoPanel.scss
@@ -5,16 +5,15 @@
   align-items: center;
   overflow: hidden; /* Masque le contenu qui dépasse la largeur de la bannière */
   white-space: nowrap; /* Empêche le texte de passer à la ligne */
-  background-color: lightblue;
+  background-color: #F8E7A2;
 
   .scrollingText {
     display: flex; 
     flex-direction: row;
-    animation: scrollText 20s linear infinite; 
+    animation: scrollText 3s linear infinite; 
 
     p {
-      color : yellow;
-      font-size: 2em;
+      font-size: 1.5em;
     }
   }
 }
